Move editing the channel features
